requirements.txt
python-docx == 0.8.10
